external_communication/requirements.txt
openai==1.12.0
supabase==2.0.0
python-dotenv==1.0.1
email-validator==2.1.0.post1   # optional but useful 
numpy==1.26.4 